requirements.txt
flask==2.3.2
openai==1.3.6
chromadb==0.4.15
sentence-transformers==2.2.2
python-dotenv==1.0.0